pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pointcam"
description = "Implementation of 'Self-supervised adversarial masking for 3D point cloud representation learning' paper."
version = "0.0.1"
requires-python = ">=3.10"
dependencies = [
    "einops==0.6.1",
    "h5py==3.8.0",
    "ipympl==0.9.3",
    "ipywidgets==8.0.4",
    "matplotlib==3.6.3",
    "numpy==1.23.5",
    "omegaconf==2.3.0",
    "opencv-python==4.7.0.72",
    "pandas==2.0.0",
    "Pillow==9.4.0",
    "plotly==5.14.1",
    "pytorch-lightning==2.0.1",
    "requests",
    "rich==13.0.0",
    "scikit-learn==1.2.1",
    "timm==0.6.13",
    "torch==2.0.1",
    "torchmetrics==0.11.4",
    "tqdm==4.65.0",
    "trimesh==3.17.1",
    "umap-learn==0.5.3",
    "wandb",
    "xformers==0.0.20",
]

[project.optional-dependencies]
3d = ["pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git"]
dev = ["black==23.3.0", "isort", "tox>=4"]

[tool.black]
line-length = 88
include_trailing_comma = true

[tool.isort]
profile = 'black'

[tool.hatch.metadata]
allow-direct-references = true
